fix: home tabs highlight
